Prevent redirects when query string changes but cache key is the same
Sometimes `redirect_canonical()` will recreate the query string and reformat it in the process. The end URL and result will be the same but the URLs do no match and so a redirect occurs. If Batcache caches this redirect then it will get the same cache key for URL formatted by `redirect_canonical()` and go into an endless redirect loop.

This change resolves the problem by preventing a redirect in the case described.